Slight core re-org.
