+++ crates/core/src/lib.rs
@@ -11,9 +11,6 @@ mod write_items;
 mod print_write_items;
 mod print;
 
-#[cfg(feature = "process")]
-pub mod process;
-
 pub mod condition_resolvers;
 pub mod conditions;
 pub mod parser_helpers;
@@ -26,8 +23,3 @@ pub use write_items::*;
 pub use print_write_items::{print_write_items, PrintWriteItemsOptions};
 pub use get_write_items::{get_write_items, GetWriteItemsOptions};
 pub use print::{print, PrintOptions};
-
-#[cfg(test)]
-mod writer_tests;
-#[cfg(test)]
-mod configuration_tests;
