+++ crates/core/src/plugins/mod.rs
@@ -1,0 +1,8 @@
+mod plugin_info;
+
+#[cfg(feature = "process")]
+pub mod process;
+#[cfg(feature = "wasm")]
+pub mod wasm;
+
+pub use plugin_info::*;
