+++ crates/core/src/plugins/process/message_processor.rs
