+++ crates/core/src/plugins/process/shared_types.rs
