+++ crates/core/src/plugins/process/stdin_out_reader_writer.rs
