+++ crates/dprint/Cargo.toml
@@ -11,7 +11,7 @@ repository = "https://github.com/dprint/dprint"
 clap = "2.33.0"
 colored = "1.9.3"
 dunce = "1.0.1"
-dprint-core = { path = "../core", version = "0.25.0", features = ["process"] }
+dprint-core = { path = "../core", version = "0.25.0", features = ["process", "wasm"] }
 dialoguer = "0.6.2"
 dissimilar = "1.0"
 globwalk = "0.8.0"
